+++ src/Bammemo.Web/Bammemo.Web.Client/BammemoComponents/NavigationMenu/BammemoSiteNav.razor
@@ -54,8 +54,6 @@
 
     protected override async Task OnInitializedAsync()
     {
-        // GlobalState.OnChange += async () => await JSRuntime.InvokeVoidAsync("bammemo.color.delayedApply");
-
         persistingSubscription = PersistentComponentState.RegisterOnPersisting(PersistData);
 
         if (!PersistentComponentState.TryTakeFromJson(PERSITING_KEY, out siteLogoUrl))
